create utilities file and constants file
