+++ tests/constants.py
@@ -1,0 +1,10 @@
+"""
+This module will store constants that will be used throughout our test suite.
+"""
+
+SERVER_PASS = "merlin-test-server"
+CERT_FILES = {
+    "ssl_cert": "test-rabbit-client-cert.pem",
+    "ssl_ca": "test-mysql-ca-cert.pem",
+    "ssl_key": "test-rabbit-client-key.pem",
+}
